refactor directory structure and python filenames
