+++ Classes/DomoticzDico.py
